+++ include/H2DE/engine/H2DE_engine.h
@@ -466,6 +466,7 @@ private:
     void _updateObjects();
 
     void _refreshObjectManagerButtons();
+    void _refreshObjectManagerCheckboxes();
     void _refreshObjectManagerInputs();
 
     void _destroy();