+++ include/H2DE/objects/H2DE_timer_object.h
@@ -196,7 +196,7 @@ public:
      * 
      * @return A map of surface names to their respective surface pointers.
      */
-    inline std::unordered_map<std::string, H2DE_Surface*> getSurfaces() const noexcept {
+    inline H2DE_SurfaceMap getSurfaces() const noexcept {
         return _surfaces; 
     }
     /**
@@ -354,7 +354,7 @@ private:
     H2DE_TimerObjectData _timerObjectData;
 
     H2DE_TextObject* _textObject = H2DE_NULL_OBJECT;
-    std::unordered_map<std::string, H2DE_Surface*> _surfaces = {};
+    H2DE_SurfaceMap _surfaces = {};
 
     H2DE_Chrono* _chrono = nullptr;