+++ src/objects/H2DE_input_object.cpp
@@ -4,9 +4,6 @@
 H2DE_InputObject::H2DE_InputObject(H2DE_Engine* e, const H2DE_ObjectData& od, const H2DE_InputObjectData& iod) : H2DE_Object(e, od), _inputObjectData(iod) {
     _inputObjectData.text.text = H2DE_TextObject::_getFormatedText(_inputObjectData.text.text);
     
-    _refreshSurfaceBuffers();
-    _refreshMaxRadius();
-
     _initCursor();
     _refreshCursor();
 }
@@ -26,8 +23,6 @@ H2DE_InputObject::~H2DE_InputObject() {
             _textObject = H2DE_NULL_OBJECT;
         }
     }
-
-    H2DE_Object::_destroySurfaces(_surfaces);
 }
 
 // ACTIONS